third_party/blink/web_tests/fast/frames/frame-with-noresize-can-be-resized-after-removal-of-noresize.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/frame-programmatic-resize.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFrames();
}

window.onload = runTests;

function runTests()
{
    if (!window.eventSender)
        return; // Cannot run this test without DRT.

    description("This tests that a frame with a source code-specified noresize attribute can be resized after programmatically removing its noresize attribute.");
    setTestFrameById("testFrame");
    shouldAllowFrameResizeAfterProcessingFrame(function(frame) { frame.removeAttribute("noresize") });
}
</script>
</head>
<frameset cols="200,*" border="10">
    <frame id="testFrame" name="testFrame" src="about:blank" noresize>
    <frame id="results" name="results" src="data:text/html,<p id='description'>This test can only be run using DRT.</p><pre id='console'></pre><script>window.onmessage = e => { if (e.data.type === 'description') { document.getElementById('description').textContent = e.data.message; } else { document.getElementById('console').appendChild(document.createTextNode(e.data.message + '\n')); } }</script>">
</frameset>
</html>
